Tests/LibWeb/Text/expected/DOM/Document-set-partial-attributes.txt

before document.fgColor=
before body attribute text=null
after document.fgColor=red
after body attribute text=red
before document.linkColor=
before body attribute link=null
after document.linkColor=red
after body attribute link=red
before document.alinkColor=
before body attribute alink=null
after document.alinkColor=red
after body attribute alink=red
before document.vlinkColor=
before body attribute vlink=null
after document.vlinkColor=red
after body attribute vlink=red
before document.bgColor=
before body attribute bgcolor=null
after document.bgColor=red
after body attribute bgcolor=red
